CDD/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
tensorflow
requests
python-dotenv
httpx